Add IsEventModeratorPerformCreateMixin to mixins in posts_comments
