+++ backend/posts_comments/mixins.py
@@ -1,3 +1,8 @@
+from rest_framework import exceptions
+
+from events.models import Role
+
+
 class VoteMixin:
     def get_vote_count(self, obj):
         return obj.votes.all().count()
@@ -10,3 +15,20 @@ class VoteMixin:
 
         votes = obj.votes.all()
         return votes.filter(user=request.user).exists()
+
+
+class IsEventModeratorPerformCreateMixin:
+    def perform_create(self, serializer):
+        event = serializer.validated_data.get('event')
+
+        try:
+            role = Role.objects.get(event=event, user=self.request.user)
+        except Role.DoesNotExist:
+            msg = "You do not have a role in related event."
+            raise exceptions.PermissionDenied(msg)
+
+        if int(role.name) < Role.NameChoice.MODERATOR:
+            msg = "You do not have permission to perform this action."
+            raise exceptions.PermissionDenied(msg)
+
+        return super().perform_create(serializer)
